Drop dbCollection songs only if it exists

+++ tasks/seed.js
@@ -1,9 +1,13 @@
 const connection = require('../config/mongoConnection');
 const songsData = require('../data/songs');
+const {songs} = require('../config/mongoCollection');
 
 async function main() {
     const db = await connection();
-    await db.collection('songs').drop();
+    let songsList = await songsData.getAll();
+    if (songsList.length !== 0) {
+        await db.collection('songs').drop();
+    }
     let song1 = await songsData.create(
         'You are loved',
         'Stars Go Dim',
